add yet another toy example, which computes the mandelbrot set by handing out work adverts to remote clients.
It is trivial parallel (master/client), uses advert for command
and data exchange.  I'm close to getting it working on the cloud
thingies...

Note that the master starts only dummy jobs at the moment, so in
order to get stuff computed, you need to start the clients
manually.  Display at the moment only works for X11, but the
output device is abstrac, so other output devices can be added.

A



git-svn-id: file://localhost/tmp/svn2git/svn@654 defb5e50-622e-49ec-a68e-d72c7db87b45