First attempt to satisfy reviewers:
1) Figure 1 adds nothing to the paper and should be removed. This will leave space for (2) and (3) 
below. 
=> removed, figure 2/3 placed below of each other

2) The second and third paragraphs of Section 2 in the original paper, now removed, should be 
retained. 
=> merged both paragraphs to one short paragraph

3) The sentence, "While SAGA represents a well-defined ... even in the presence of failures", found 
in Section 3 of the original, now removed, should be retained. 
=> added to introduction

4) Care needs to exercised in the use of commas throughout. For example on p9, "While it has 
been shown, that resources ..." the comma should be removed. 
=> done. 

5) Care needs to be exercised with regard to spelling throughout. For example: on p1, "is an 
important example of a first-generation Grid applications", this should be "application"; "hepatatis" 
should be "hepatitis"; on p5, "must initializes", should be, "must initialize". 
=> done.

6) On p2, "We provide irrefutable evidence ...". Evidence has been provided only for the cases 
studied. There is insufficient rigour in the analysis to merit the use of "irrefutable". Please remove, 
"irrefutable". 
=> done.

That makes in total: 10 1/3 pages

Now they question left is: What do we do about the request for more in section 2 and 6?

Andre




git-svn-id: file://localhost/tmp/svn2git/svn@841 defb5e50-622e-49ec-a68e-d72c7db87b45