Soooo, the new digedag is now working, and should be much more stable and scalable than the old code.  The structure remained mostly the same, but the API was simplified somewhat - documentation is/should be forthcoming.
The implemented default scheduler maintains sepatare queues for
tasks and edges, and runs up to 10 tasks (hardcoded) in parallel
for each queue.  That seems to scale fine for up to 100 tasks,
where we then hit resource shortage (# of threads and/or open
files).  That number will depend on the backend used, and should
be significantly larger for adaptors with low local resource
requirements.

Most pointers are now shared pointers, exept that the scheduler
maintains a pointer to the DAG.  But as the scheduler is created
and destroyed by the DAG, that should be ok.

Next steps:  
  - tests with larged DAGs, to make Kate happy
  - simplify scheduler exchange and extension, to make Saurabh
    happy
  - documentation, to make everybody happy

Further simplification of the code can be expected once the
saga::task_container behaves as it should.

A


git-svn-id: file://localhost/tmp/svn2git/svn@2034 defb5e50-622e-49ec-a68e-d72c7db87b45