adding an example plot with error bars.
We don't have stdev for the HC, so thta is somewhat confusing.
Also, the stdev is indeed in the order of the point size for
most data points, so are hard to see on the plots, even in
different colors.  So, I guess that it would be best to leave
them out of the graphs after all, and to discuss stdevs in the
text.  Please let me know if you think otherwise...

A.


git-svn-id: file://localhost/tmp/svn2git/svn@3043 defb5e50-622e-49ec-a68e-d72c7db87b45