major cleanup:
  - consistent job numbering
  - master now completely configurable via ini
    - output devices
    - location of output picture
    - mandelbrot parameters
    - job item decomposition
    - advert dir location (base)
    - use session id instead of pid, to avoid id collisions
  - master is somewhat resilient against 
    - missing endoints
    - failing endpoints
    - missing clients
    - stalling clients
    - misbehaving clients
    - bad clients
    - incomplete work
    - corrupted adverts
  - clients accept additional commands to mimic
    stalling/misbehaving/bad clients
  - clients will discontinue if they idle for too long, to avoid
    client zombies
  - master output is now actually readable

Note that this commit is, as far as I can tell, backward
compatible to the old code: old clients will work with the new
masterm, and vice versa.  The next commit will likely *break*
that.  

Note to the students playing with Mandelbrot right now: "Don't
fix it if it ain't broken!" - i.e.  only upgrade if you know
what you are doing.  

Cheers, Andre.



git-svn-id: file://localhost/tmp/svn2git/svn@3305 defb5e50-622e-49ec-a68e-d72c7db87b45