Updated TOC and assigned chapters.
I think the Abstract reads ok. One thing we should put some focus on (and maybe mention in the abstract), is a comprehensive study of related work.  IMHO this would be a very important aspect in "THE" SAGA paper and would help to justify not only SAGA's existence, but also emphasize its importance in a changing dist. comp. landscape that is becoming more and more globalised.

Any Thoughts?



git-svn-id: file://localhost/tmp/svn2git/svn@3311 defb5e50-622e-49ec-a68e-d72c7db87b45