unlock (IV)
Bluntly, the more I look at the pilot API, the less I like its
details.  Having said that, section 4 looks coherent, but is
somewhat dense (many names).  A simple and short scenario
walk-through might help to clear this up for the reader, but
I am not sure if we have the space for that.  What is the page
limit actually?

Cheers, Andre.



git-svn-id: file://localhost/tmp/svn2git/svn@7234 defb5e50-622e-49ec-a68e-d72c7db87b45